drop needless chmods
The default value for the $permissions argument to mkdir is already 0777,
so no need to mkdir here. The only case it would matter would be if the
directory was manually created, but with the wrong permissions. But in
that case, it might be an intentional choice, so let's not meddle with
it.